+++ client/src/modules/extract-quotes.ts
