+++ client/src/modules/state.ts
@@ -1,4 +1,5 @@
 import { autorun, makeAutoObservable, observable, reaction } from "mobx";
+import { Uri } from 'vscode';
 import { client } from '../extension';
 
 class ExtensionStateStore {
@@ -6,6 +7,9 @@ class ExtensionStateStore {
 	longProcessing = false;
 	diagnosing = false;
 	preprocessing = false;
+
+	isUsingAdv3Lite = false
+	chosenMakefileUri: Uri | undefined;
 
 	constructor() {
 		makeAutoObservable(this, {
@@ -33,6 +37,19 @@ class ExtensionStateStore {
 	
 	isPreprocessing() { return this.preprocessing; }
 
+	setUsingAdv3LiteStatus(state: boolean) { this.isUsingAdv3Lite = state }
+
+	getUsingAdv3LiteStatus() {return this.isUsingAdv3Lite; }
+	
+	getChosenMakefileUri(): Uri | undefined {
+		return this.chosenMakefileUri;
+	}
+
+	setChosenMakefileUri(chosenMakefileUri :Uri | undefined) {
+		this.chosenMakefileUri = chosenMakefileUri;
+	}
+
+	
 }
 
 export const extensionState = new ExtensionStateStore();
